Ders Silme Islemi

+++ C#_Form_ile_Gorsel_ve_OOP/C#_Form_ile_Gorsel_ve_OOP_501/KatmanliMimariProje/EntityLayer/EntityDers.cs
@@ -8,7 +8,7 @@ namespace EntityLayer
 {
 	public class EntityDers
 	{
-		public int DersID { get; set; }
+		public byte DersID { get; set; }
 		public string DersAd { get; set; }
 	}
 }
